lambda functoin for verifying videos
